+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Fixes
+* Bump Wallet availability wait time to 6m (was 3m) because when your CPU is heavily-loaded (e.g. compiling something) it can take longer to compile the Wallet
 
 # 0.5.3
 ### Fixes